backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
stripe
razorpay
httpx
pydantic
pydantic-settings
email-validator
google-generativeai
google-genai